Revert "🚀 Release v0.0.3"
