+++ apps/api/src/user/user.service.ts
@@ -15,10 +15,4 @@ export class UserService {
     });
     return user;
   }
-
-  public async delete(id: string) {
-    const user = new User();
-    user.id = id;
-    return await this.dataSource.manager.delete(User, { id });
-  }
 }
