+++ apps/api/test/e2e/jest-e2e.json
@@ -8,7 +8,8 @@
   },
   "roots": ["<rootDir>/apps/", "<rootDir>/libs/"],
   "moduleNameMapper": {
-    "^@libs(|/.*)$": "<rootDir>/libs/src/$1",
+    "^@libs/typeorm(|/.*)$": "<rootDir>/libs/typeorm/src/$1",
+    "^@libs/utils(|/.*)$": "<rootDir>/libs/utils/src/$1",
     "^@app/test(|/.*)$": "<rootDir>/libs/test/src/$1",
     "^@api/test(|/.*)$": "<rootDir>/apps/api/test/$1",
     "^@api/src(|/.*)$": "<rootDir>/apps/api/src/$1"
