+++ apps/api/test/unit/user/user.service.unit-spec.ts
@@ -1,14 +1,13 @@
 import { UserRepository } from '@api/src/user/user.repository';
 import { UserService } from '@api/src/user/user.service';
+import { DataSoureMock } from '@libs/typeorm/utils/test/mock.test-util';
 import { Test, TestingModule } from '@nestjs/testing';
 import { DataSource } from 'typeorm';
-import { type DeepMockProxy, mockDeep } from 'jest-mock-extended';
 
 describe('[UserService]', () => {
   let userService: UserService;
   let userRepository: UserRepository;
   let module: TestingModule;
-  let dataSoure: DeepMockProxy<DataSource>;
 
   beforeEach(async () => {
     module = await Test.createTestingModule({
@@ -17,7 +16,7 @@ describe('[UserService]', () => {
         UserRepository,
         {
           provide: DataSource,
-          useValue: mockDeep<DataSource>(),
+          useClass: DataSoureMock,
         },
       ],
     }).compile();
@@ -30,17 +29,13 @@ describe('[UserService]', () => {
     it('user create의 결과로 반환되는 User객체의 id는 주어진 id와 일치해야 한다. ', async () => {
       const a = '3';
 
+      const spyFn = jest.spyOn(DataSoureMock.prototype, 'transaction');
+
       expect((await userService.create(a)).id).toBe(a);
+
+      expect(spyFn).toBeCalled();
     });
   });
-
-  //   describe('delete()', () => {
-  //     it('user delete시 transaction이 호출되어야 한다. ', async () => {
-  //       const id = '3';
-
-  //       expect(await userService.delete(id)).toBeInstanceOf(Promise<DeleteResult>);
-  //     });
-  //   });
 
   afterAll(async () => {
     if (module) {
