--- libs/src/typeorm/domain/default-entity.entity.ts
@@ -1,17 +1,0 @@
-import { PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn, DeleteDateColumn } from 'typeorm';
-
-export abstract class DefaultEntity {
-  @CreateDateColumn()
-  created_at: Date;
-
-  @UpdateDateColumn()
-  updated_at: Date;
-
-  @DeleteDateColumn()
-  deletedAt!: Date | null;
-}
-
-export abstract class DefaultAutoIdEntity extends DefaultEntity {
-  @PrimaryGeneratedColumn({ unsigned: true })
-  id: number;
-}
