--- libs/src/typeorm/seed/user.seeder.ts
@@ -1,11 +1,0 @@
-import { Seeder, SeederFactoryManager } from 'typeorm-extension';
-import { DataSource } from 'typeorm';
-import { User } from '@libs/typeorm/domain/user/user.entity';
-
-export default class UserSeeder implements Seeder {
-  public async run(dataSource: DataSource, factoryManager: SeederFactoryManager): Promise<any> {
-    const userFactory = await factoryManager.get(User);
-
-    await userFactory.saveMany(30);
-  }
-}
