+++ libs/typeorm/src/domain/default-entity.entity.ts
