+++ libs/typeorm/src/utils/test/mock.test-util.ts
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/no-empty-function */
+export class DataSoureMock {
+  async transaction(): Promise<void> {}
+}
