+++ package.json
@@ -17,13 +17,11 @@
     "test:api:unit": "jest --config ./apps/api/test/unit/jest-e2e.json --runInBand",
     "test:api:unit:cov": "jest --config ./apps/api/test/unit/jest-e2e.json --ci --json --coverage --testLocationInResults --outputFile=report.json --reporters=default --reporters=jest-junit --runInBand",
     "test:api:e2e": "jest --config ./apps/api/test/e2e/jest-e2e.json --runInBand",
-    "test:api:e2e:cov": "jest --config ./apps/api/test/e2e/jest-e2e.json --ci --json --coverage --testLocationInResults --outputFile=report.json --reporters=default --reporters=jest-junit --runInBand",
+    "test:api:e2e:cov": "jest --config ./apps/api/test/jest-e2e.json --ci --json --coverage --testLocationInResults --outputFile=report.json --reporters=default --reporters=jest-junit --runInBand",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-    "wait-db": "bash ./wait-db.sh",
-    "preseed":"yarn build",
-    "seed:local": "export NODE_ENV=local && ts-node ./node_modules/typeorm-extension/bin/cli.cjs seed:run -d dist/libs/typeorm/src/seed/data-source.js"
+    "wait-db": "bash ./wait-db.sh"
   },
   "jest-junit": {
     "outputName": "jest-junit.xml",
@@ -63,7 +61,6 @@
     "eslint-plugin-prettier": "^4.0.0",
     "jest": "28.0.3",
     "jest-junit": "^16.0.0",
-    "jest-mock-extended": "^3.0.6",
     "prettier": "^2.3.2",
     "source-map-support": "^0.5.20",
     "supertest": "^6.1.3",
@@ -71,7 +68,6 @@
     "ts-loader": "^9.2.3",
     "ts-node": "^10.0.0",
     "tsconfig-paths": "4.0.0",
-    "typeorm-extension": "^3.5.1",
     "typescript": "^4.3.5"
   },
   "jest": {
